feat: Quran learning packages, Smart Khatma, and design-system UI migration

+++ apps/tilawa/lib/features/home/presentation/widgets/home_hero_collapsed_bar.dart
@@ -77,7 +77,7 @@ class HomeHeroCollapsedBar extends StatelessWidget {
                       chromeAlpha,
                 ),
                 offset: tokens.shadowOffsetSmall,
-                blurRadius: tokens.spaceSmall.toDouble(),
+                blurRadius: tokens.spaceSmall,
               ),
             ]
           : null,
