+++ apps/tilawa/lib/features/share/presentation/utils/selected_quran_range_page.dart
@@ -301,5 +301,5 @@ double _scaledGap(
   required double min,
   required double max,
 }) {
-  return (lineHeight * factor).clamp(min, max).toDouble();
+  return (lineHeight * factor).clamp(min, max);
 }
